twitter_bot/requirements.txt
python-dotenv
tweepy
feedparser
APScheduler
